+++ examples/basic/usage/simple.php
@@ -1,0 +1,11 @@
+<?php
+
+use Castor\Attribute\AsTask;
+
+use function Castor\io;
+
+#[AsTask()]
+function hello(): void
+{
+    io()->writeln('Hello from castor.');
+}